+++ client/model/index.ts
@@ -1,6 +1,6 @@
 import { BASE_PATH } from "@env";
 import axios from "axios";
 
-export const api = axios.create({
-  baseURL: BASE_PATH,
-});
+// export const api = axios.create({
+//   baseURL: BASE_PATH,
+// });
